+++ custom_components/unifi_insights/api/network/models/client.py
@@ -19,6 +19,7 @@ class ClientType(str, Enum):
     WIRED = "WIRED"
     WIRELESS = "WIRELESS"
     VPN = "VPN"
+    TELEPORT = "TELEPORT"
 
     @classmethod
     def _missing_(cls, value: object) -> ClientType | None: